Add exchange ratio function, remove six and clean code
