+++ tests/test_money.py
@@ -8,7 +8,7 @@ import pytest
 from txmoney.money.exceptions import (
     CurrencyDoesNotExist, CurrencyMismatch, IncorrectMoneyInputError
 )
-from txmoney.money.models.models import CURRENCIES, Currency, Money
+from txmoney.money.models.money import CURRENCIES, Currency, Money
 from txmoney.settings import txmoney_settings as settings
