+++ txmoney/money/__init__.py
@@ -1,7 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import
-
-from .models.models import Currency, Money  # noqa
-
+from __future__ import unicode_literals
 
 __version__ = '0.6.0'
