+++ txmoney/money/models/utils.py
@@ -1,7 +1,10 @@
+# coding=utf-8
+from __future__ import absolute_import, unicode_literals
+
 from django.db.models import F
 from django.db.models.expressions import BaseExpression
 
-from txmoney.money.models.models import Money
+from .money import Money
 
 
 def get_currency_field_name(name):
