+++ txmoney/rates/backends.py
@@ -11,9 +11,9 @@ from django.utils.six import iteritems, with_metaclass
 
 from ..compat import urlopen
 from ..settings import txmoney_settings as settings
-from ..utils import parse_rates_to_base_currency
 from .exceptions import TXRateBackendError
 from .models import Rate, RateSource
+from .utils import parse_rates_to_base_currency
 
 
 class BaseRateBackend(with_metaclass(ABCMeta)):
