+++ txmoney/rates/utils.py
@@ -1,12 +1,29 @@
 # coding=utf-8
 from __future__ import absolute_import, division, unicode_literals
 
+from datetime import date
 from decimal import Decimal
 
 from django.utils.six import iteritems
 
-from txmoney.money.models.models import CURRENCIES
-from txmoney.settings import txmoney_settings as settings
+from ..settings import txmoney_settings as settings
+from .models import Rate
+
+
+def exchange_ratio(currency_from, currency_to, ratio_date=None):
+    """
+    Return exchange ratio between two currencies for a date
+    """
+    ratio_date = ratio_date or date.today()
+    rate_from = rate_to = Decimal(1)
+
+    if currency_from != currency_to:
+        if currency_from != settings.BASE_CURRENCY:
+            rate_from = Rate.objects.get_for_date(currency_from, ratio_date).value
+        if currency_to != settings.BASE_CURRENCY:
+            rate_to = Rate.objects.get_for_date(currency_to, ratio_date).value
+
+    return rate_to / rate_from
 
 
 def parse_rates_to_base_currency(rates, origin_currency):
@@ -25,19 +42,5 @@ def parse_rates_to_base_currency(rates, origin_currency):
     for currency, value in iteritems(rates):
         rates[currency] = value * rate
     rates[origin_currency] = rate
-
-    return rates
-
-
-def clean_rates(rates):
-    """
-    Clean rates list to accept only system supported rates
-    :param rates: rates currencies
-    """
-    intersection_rates = set(CURRENCIES).intersection(set(rates))
-
-    for key in rates.keys():
-        if key not in intersection_rates:
-            del rates[key]
 
     return rates
